Figures formatted nicely, lots of other fixes.
In particular, I went through all of the TODOs and patched them up,
except for those that will require more thinking/work/or input.
For instance, the montages are still not complete, but I elaborated
the todo's with figure references so that other readers won't be
weirded out.

Also, I fixed a bunch of references, and put in citet references
where necessary.

Lastly, in addition to many wording changes and spelling/grammar
fixes, I added in errorbars on the ADNI-Xval plot, changed plot
text and point sizes, and included p/t-values for Pearson
correlations mentioned in the text.  w00t.

I'm going to send this off to Mallar now.